Struct brainfk::bf::cell::Cell

pub struct Cell { /* private fields */ }

Memory cell This is where the cell is created and managed by bf. visualizaion… |—|—|—|—|—|—|—|—| | 0 | 0 | 0 | 0 | 0 | 0 | 0 | 0 | |—|—|—|—|—|—|—|—| ^
| ptr

Implementations

impl Cell

pub fn new() -> Self

creates a new memory cell

pub fn len(&self) -> usize

returns the size of memory cell which is 30000 by default. |—|—|—|—|—|—|—|—| | 0 | 0 | 0 | 0 | 0 | 0 | 0 | 0 | |—|—|—|—|—|—|—|—| ^ * len = 8 | ptr

pub fn inc(&mut self) -> CellResult

increases the current cell by one.

|—|—|—|—|—|—|—|—| | 1 | 0 | 0 | 0 | 0 | 0 | 0 | 0 | |—|—|—|—|—|—|—|—| ^
| ptr

pub fn dec(&mut self) -> CellResult

decreases the current cell by one. |—|—|—|—|—|—|—|—| | 0 | 0 | 0 | 0 | 0 | 0 | 0 | 0 | |—|—|—|—|—|—|—|—| ^
| ptr

pub fn ptr(&self) -> usize

returns the pointer to the current cell |—|—|—|—|—|—|—|—| | 0 | 0 | 0 | 0 | 0 | 0 | 0 | 0 | |—|—|—|—|—|—|—|—| ^
| *ptr

pub fn next(&mut self) -> CellResult

|—|—|—|—|—|—|—|—| | 0 | 0 | 0 | 0 | 0 | 0 | 0 | 0 | |—|—|—|—|—|—|—|—| ^
| *ptr

pub fn previous(&mut self) -> CellResult

rewinds to the previous cell. While the pointer is on cell 0, trying to rewind will cause an error, please beware of that. |—|—|—|—|—|—|—|—| | 0 | 0 | 0 | 0 | 0 | 0 | 0 | 0 | |—|—|—|—|—|—|—|—| ^
| *ptr

pub fn data(&self) -> i64

pub fn set(&mut self, b: i64)

pub fn cells<'a>(&'a self) -> &'a Box<[i64]>